+++ test/Elements.Tests/ModelTests.cs
@@ -4,6 +4,10 @@ using Xunit;
 using Elements.Geometry;
 using Elements.Serialization.glTF;
 using System.Collections.Generic;
+using Elements.Generate;
+using Newtonsoft.Json;
+using Elements.Geometry.Solids;
+using System.Linq;
 
 namespace Elements.Tests
 {
@@ -60,7 +64,106 @@ namespace Elements.Tests
             var errors = new List<string>();
             var model = Model.FromJson(modelStr, errors);
             Assert.Equal(2, model.Elements.Count);
-            Assert.Equal(1, errors.Count);
+            Assert.Equal(2, errors.Count);
+        }
+
+        /// <summary>
+        /// Test whether two models, containing user defined types, can be 
+        /// deserialized and merged into one model.
+        /// </summary>
+        [Fact(Skip="ModelMerging")]
+        public void MergesModelsWithUserDefinedTypes()
+        {
+            var schemas = new[]{
+                "../../../models/Merge/Envelope.json",
+                "../../../models/Merge/FacadePanel.json",
+                "../../../models/Merge/Level.json"
+            };
+
+            var asm = TypeGenerator.GenerateInMemoryAssemblyFromUrisAndLoad(schemas);
+            var facadePanelType = asm.GetType("Elements.FacadePanel");
+            Assert.NotNull(facadePanelType);
+            var envelopeType = asm.GetType("Elements.Envelope");
+            Assert.NotNull(envelopeType);
+            var model1 = JsonConvert.DeserializeObject<Model>(File.ReadAllText("../../../models/Merge/facade.json"));
+            var count1 = model1.Elements.Count;
+
+            var model2 = JsonConvert.DeserializeObject<Model>(File.ReadAllText("../../../models/Merge/structure.json"));
+            var count2 = model2.Elements.Count;
+
+            var merge = new Model();
+            merge.AddElements(model1.Elements.Values);
+            merge.AddElements(model2.Elements.Values);
+            merge.ToGlTF("models/Merge.glb");
+        }
+
+        [Fact]
+        public void ElementWithDeeplyNestedElementSerializesCorrectly()
+        {
+            var p = new Profile(Polygon.Rectangle(1,1));
+            // Create a mass overiding its representation.
+            // This will introduce a profile into the serialization for the
+            // representation. This should be serialized correctly.
+            var mass1 = new Mass(p,
+                                1,
+                                BuiltInMaterials.Mass,
+                                new Transform(),
+                                new Representation(new List<SolidOperation> { new Extrude(p, 2, Vector3.ZAxis, 0, false) }));
+            // A second mass that uses a separate embedded profile.
+            // This is really a mistake because the user wants the profile
+            // that they supply in the constructor to be used, but the profile
+            // supplied in the representation will override it.
+            var mass2 = new Mass(p,
+                                1,
+                                BuiltInMaterials.Mass,
+                                new Transform(),
+                                new Representation(new List<SolidOperation> { new Extrude(new Profile(Polygon.Rectangle(1,1)), 2, Vector3.ZAxis, 0, false) }));
+            var model = new Model();
+            model.AddElement(mass1);
+            model.AddElement(mass2);
+            Assert.Equal(2, model.AllElementsOfType<Profile>().Count());
+            Assert.Equal(2, model.AllElementsOfType<Mass>().Count());
+            Assert.Equal(1, model.AllElementsOfType<Material>().Count());
+
+            var json = model.ToJson();
+            File.WriteAllText("./deepSerialize.json", json);
+
+            var newModel = Model.FromJson(json);
+            Assert.Equal(2, newModel.AllElementsOfType<Profile>().Count());
+            Assert.Equal(2, newModel.AllElementsOfType<Mass>().Count());
+            Assert.Equal(1, newModel.AllElementsOfType<Material>().Count());
+        }
+
+        [Fact]
+        public void CoreElementTransformsAreIdempotentDuringSerialization()
+        {
+            var model = new Model();
+            // Create a floor with an elevation.
+            // This will mutate the element's transform.
+            var floor = new Floor(Polygon.L(20, 20, 5), 0.1, new Transform(0,0,0.5));
+            model.AddElement(floor);
+
+            var beam = new Beam(new Line(Vector3.Origin, new Vector3(5,5,5)), Polygon.Rectangle(0.1,0.1));
+            model.AddElement(beam);
+
+            var wall = new StandardWall(new Line(Vector3.Origin, new Vector3(20,0,0)), 0.2, 3.5);
+            model.AddElement(wall);
+
+            // Serialize the floor, recording the mutated transform.
+            var json = model.ToJson();
+
+            // Deserialize the floor, which will deserialize the elevation
+            // and add it to the transform.
+            var newModel = Model.FromJson(json);
+            
+            var newFloor = newModel.AllElementsOfType<Floor>().First();
+            Assert.Equal(floor.Transform, newFloor.Transform);
+
+            var newBeam = newModel.AllElementsOfType<Beam>().First();
+            Assert.Equal(beam.Transform, newBeam.Transform);
+
+            var newWall = newModel.AllElementsOfType<StandardWall>().First();
+            Assert.Equal(wall.Transform, newWall.Transform);
         }
   
         private Model QuadPanelModel()