+++ test/Elements.Tests/StructuralFramingTests.cs
@@ -121,5 +121,28 @@ namespace Elements.Tests
             Assert.Equal(line, brace.Curve);
             this.Model.AddElement(brace);
         }
+
+        [Fact]
+        public void Setbacks()
+        {
+            this.Name = "BeamSetbacks";
+            var line = new Line(Vector3.Origin, new Vector3(3,3,0));
+            var mc = new ModelCurve(line, BuiltInMaterials.XAxis);
+            this.Model.AddElement(mc);
+            // Normal setbacks
+            var beam = new Beam(line, this._testProfile, BuiltInMaterials.Steel, 2.0, 2.0);
+            this.Model.AddElement(beam);
+
+            var line1 = new Line(new Vector3(2,0,0), new Vector3(5,3,0));
+            var mc1 = new ModelCurve(line1, BuiltInMaterials.XAxis);
+            this.Model.AddElement(mc1);
+
+            var sb = line1.Length()/2;
+            // Setbacks longer in total than the beam.
+            // We are testing to ensure that the beam gets created
+            // without throwing. It will not have setbacks.
+            var beam1 = new Beam(line1, this._testProfile, BuiltInMaterials.Steel, sb, sb);
+            this.Model.AddElement(beam1);
+        }
     }
 }